Renamed fourierencoding and trajectorypreparation
