+++ src/xGadgetron/cGadgetron/include/sirf/Gadgetron/FourierEncoding.h
