+++ src/xGadgetron/cGadgetron/include/sirf/Gadgetron/TrajectoryPreparation.h
