+++ src/xGadgetron/cGadgetron/tests/mrtests.cpp
@@ -43,8 +43,8 @@ limitations under the License.
 #include "sirf/Gadgetron/gadgetron_data_containers.h"
 #include "sirf/Gadgetron/gadgetron_x.h"
 
-#include "sirf/Gadgetron/fourierencoding.h"
-#include "sirf/Gadgetron/trajectorypreparation.h"
+#include "sirf/Gadgetron/FourierEncoding.h"
+#include "sirf/Gadgetron/TrajectoryPreparation.h"
 
 #include "mrtest_auxiliary_funs.h"
